requirements.txt
flask
psycopg2
google-generativeai
flask-cors